MM-57: Add UT and IT

+++ src/main/java/ii/cipriantarlev/marketmanagementapi/plu/PluRepository.java
@@ -5,7 +5,7 @@ package ii.cipriantarlev.marketmanagementapi.plu;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PluRepository extends JpaRepository<Plu, Integer> {
+public interface PluRepository extends JpaRepository<Plu, Long> {
 
 	Plu findFirst1ByOrderByValueDesc();
 }
